refactor(supernova): use of reactQuery states instead of zustand
