+++ apps/supernova/src/AppContent.jsx
@@ -5,15 +5,8 @@
 
 import React from "react"
 import { useActions, Messages } from "@cloudoperators/juno-messages-provider"
-import { Container } from "@cloudoperators/juno-ui-components"
-import {
-  useAlertsIsUpdating,
-  useAlertsUpdatedAt,
-  useAlertsTotalCounts,
-  useSilencesError,
-  useGlobalsActiveSelectedTab,
-  useAlertsError,
-} from "./components/StoreProvider"
+import { Container, Stack, Spinner } from "@cloudoperators/juno-ui-components"
+import { useAlertsUpdatedAt, useAlertsTotalCounts, useGlobalsActiveSelectedTab } from "./components/StoreProvider"
 import AlertsList from "./components/alerts/AlertsList"
 import RegionsList from "./components/regions/RegionsList"
 import StatusBar from "./components/status/StatusBar"
@@ -23,34 +16,32 @@ import AlertDetail from "./components/alerts/AlertDetail"
 import PredefinedFilters from "./components/filters/PredefinedFilters"
 import SilencesList from "./components/silences/SilencesList"
 
+import { useBoundQuery } from "./hooks/useBoundQuery"
+
 const AppContent = () => {
   const { addMessage } = useActions()
 
   // alerts
   const totalCounts = useAlertsTotalCounts()
-  const isAlertsUpdating = useAlertsIsUpdating()
   const updatedAt = useAlertsUpdatedAt()
-
-  // silences
-  const silencesError = useSilencesError()
-  const alertsError = useAlertsError()
 
   const activeSelectedTab = useGlobalsActiveSelectedTab()
 
+  const { error: alertsError, isLoading: isAlertsLoading } = useBoundQuery("alerts")
+  const { error: silencesError, isLoading: isSilencesLoading } = useBoundQuery("silences")
   // since the API call is done in a web worker and not logging aware, we need to show the error just in case the user is logged in
   if (silencesError) {
     addMessage({
       variant: "error",
-      text: parseError(silencesError),
+      text: parseError(alertsError),
     })
   }
 
   // since the API call is done in a web worker and not logging aware, we need to show the error just in case the user is logged in
-
   if (alertsError) {
     addMessage({
       variant: "error",
-      text: parseError(alertsError),
+      text: parseError(silencesError),
     })
   }
 
@@ -62,15 +53,33 @@ const AppContent = () => {
         <>
           <AlertDetail />
           <RegionsList />
-          <>
-            <PredefinedFilters />
-            <Filters />
-            <StatusBar totalCounts={totalCounts} isUpdating={isAlertsUpdating} updatedAt={updatedAt} />
-            <AlertsList />
-          </>
+          {isAlertsLoading ? (
+            <Stack gap="2">
+              <span>Loading</span>
+              <Spinner variant="primary" />
+            </Stack>
+          ) : (
+            <>
+              <PredefinedFilters />
+              <Filters />
+              <StatusBar totalCounts={totalCounts} isUpdating={isAlertsLoading} updatedAt={updatedAt} />
+              <AlertsList />
+            </>
+          )}
         </>
       )}
-      {activeSelectedTab === "silences" && <SilencesList />}
+      {activeSelectedTab === "silences" && (
+        <>
+          {isSilencesLoading ? (
+            <Stack gap="2">
+              <span>Loading</span>
+              <Spinner variant="primary" />
+            </Stack>
+          ) : (
+            <SilencesList />
+          )}
+        </>
+      )}
     </Container>
   )
 }
